Task 04 (phase 3): extract :core:network, :core:database, :core:datastore
Move the network layer (ArduinoApi, ArduinoResponse, NetworkModule), the Room
layer (AppDatabase, dao/*, entity/*) and DataStoreManager into three new leaf
modules. Rewrite packages and imports across all consumers and wire the new
deps into each module that uses them.

DI note: CoreComponent's @Component(modules=[NetworkModule, AppModule]) plus its
provisions (AppDatabase, DataStoreManager, ArduinoApi, NavigationDispatcher) are
part of :core's public API, so downstream Dagger components (BaseComponent and
the feature components depending on CoreComponent) must see those types. :core
therefore re-exports :core:{common,model,network,database,datastore} via `api`
as an intermediate bridge; this is right-sized away once consumers depend on the
specific modules directly (later phase). Also added the missing explicit
NetworkModule import to CoreComponent (was a same-package reference).

Builds green: assembleDebug + detektDebug + ktlintCheck.
