+++ base/src/main/java/com/timhome/base/BaseModule.kt
@@ -1,6 +1,6 @@
 package com.timhome.base
 
-import com.timhome.core.data.DataStoreManager
+import com.timhome.core.datastore.DataStoreManager
 import com.timhome.core.data.repository.ArduinoRepository
 import com.timhome.core.data.repository.ArduinoRepositoryImpl
 import com.timhome.core.common.FeatureScope
