+++ core/database/build.gradle.kts
@@ -1,0 +1,12 @@
+plugins {
+    id("timHome.android.library")
+    id("timHome.quality.convention.plugin")
+}
+
+android {
+    namespace = "com.timhome.core.database"
+}
+
+dependencies {
+    implementation(project(":core:common"))
+}
