+++ core/database/src/main/java/com/timhome/core/database/dao/UserDao.kt
@@ -1,9 +1,9 @@
-package com.timhome.core.data.db.dao
+package com.timhome.core.database.dao
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
-import com.timhome.core.data.db.entity.UserEntity
+import com.timhome.core.database.entity.UserEntity
 
 @Dao
 interface UserDao {
