+++ core/database/src/main/java/com/timhome/core/database/entity/UserEntity.kt
@@ -1,4 +1,4 @@
-package com.timhome.core.data.db.entity
+package com.timhome.core.database.entity
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
