+++ core/datastore/build.gradle.kts
@@ -1,0 +1,12 @@
+plugins {
+    id("timHome.android.library")
+    id("timHome.quality.convention.plugin")
+}
+
+android {
+    namespace = "com.timhome.core.datastore"
+}
+
+dependencies {
+    implementation(libs.datastore)
+}
