+++ core/network/src/main/java/com/timhome/core/network/api/response/ArduinoResponse.kt
@@ -1,3 +1,3 @@
-package com.timhome.core.data.api.response
+package com.timhome.core.network.api.response
 
 data class ArduinoResponse(val temperature: Double, val co2: Int)
