+++ feature/home/src/main/java/com/timhome/home/data/repository/WeatherRepository.kt
@@ -1,6 +1,6 @@
 package com.timhome.home.data.repository
 
-import com.timhome.core.data.DataStoreManager
+import com.timhome.core.datastore.DataStoreManager
 import com.timhome.core.common.CallStatus
 import com.timhome.core.common.Constant.CODE_200
 import com.timhome.home.data.api.WeatherApi
