+++ mock/build.gradle.kts
@@ -10,6 +10,8 @@ android {
 dependencies {
     implementation(project(":core"))
     implementation(project(":core:common"))
+    implementation(project(":core:database"))
+    implementation(project(":core:network"))
     implementation(libs.bundles.room)
     implementation(libs.dagger)
     ksp(libs.dagger.compiler)
