+++ settings.gradle.kts
@@ -25,6 +25,9 @@ include(":app")
 include(":core")
 include(":core:common")
 include(":core:model")
+include(":core:network")
+include(":core:database")
+include(":core:datastore")
 include(":benchmark")
 include(":mock")
 include(":base")
